Add files via upload
This is my map of Bay Ridge. It includes the yellow line representing the R train route, the orange line representing the B16, the green line representing the B37, and the blue line representing the B63. The locations marked off are Fort Hamilton High School, my house, and the Fort Hamilton branch of the Brooklyn Public Library.